Полностью заменил uart_stdio на uart-debug, SPI теперь работает и на ATmega328p, и на ATmega128.

+++ granum-space/src/i2c.c
@@ -46,7 +46,7 @@ i2c_error_t i2c_status_to_error(uint8_t status)
 
 void i2c_init()
 {
-	TWBR = (F_CPU/F_SCL - 16)/2/1;
+	TWBR = (uint8_t)(F_CPU/F_SCL - 16)/2/1;
 }
